pyproject.toml
[tool.poetry]
name = "setup-teardown"
version = "0.1.0"
description = ""
authors = ["ges <gerrit.schrader@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^6.0.2"
black = "^20.8b1"
mypy = "^0.782"
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 100
